Package rename settheory -> sets, class rename SetTheory -> SetOperation
