+++ adam-core/src/main/scala/org/bdgenomics/adam/rdd/sets/Closest.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.rdd.settheory
+package org.bdgenomics.adam.rdd.sets
 
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.models.ReferenceRegion
@@ -32,7 +32,7 @@ import scala.reflect.ClassTag
  * @tparam RX The resulting type of the right after the operation.
  */
 sealed trait Closest[T, U <: GenomicRDD[T, U], X, Y <: GenomicRDD[X, Y], RT, RX]
-    extends SetTheoryBetweenCollections[T, U, X, Y, RT, RX] {
+    extends SetOperationBetweenCollections[T, U, X, Y, RT, RX] {
 
   override protected def condition(firstRegion: ReferenceRegion,
                                    secondRegion: ReferenceRegion,
@@ -202,7 +202,7 @@ case class ShuffleClosestRegion[T, U <: GenomicRDD[T, U], X, Y <: GenomicRDD[X, 
   protected val threshold: Long = Long.MaxValue,
   protected val optPartitions: Option[Int] = None)
     extends Closest[T, U, X, Y, T, Iterable[X]]
-    with VictimlessSetTheoryBetweenCollections[T, U, X, Y, T, Iterable[X]] {
+    with VictimlessSetOperationBetweenCollections[T, U, X, Y, T, Iterable[X]] {
 
   override protected def emptyFn(left: Iterator[(ReferenceRegion, T)],
                                  right: Iterator[(ReferenceRegion, X)]): Iterator[(T, Iterable[X])] = {
