+++ adam-core/src/test/scala/org/bdgenomics/adam/rdd/sets/ClosestSuite.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.rdd.settheory
+package org.bdgenomics.adam.rdd.sets
 
 import org.bdgenomics.adam.models.ReferenceRegion
 import org.bdgenomics.adam.util.ADAMFunSuite
