+++ adam-core/src/test/scala/org/bdgenomics/adam/rdd/sets/LeftOuterShuffleRegionJoinSuite.scala
@@ -15,19 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.rdd
+package org.bdgenomics.adam.rdd.sets
 
 import org.apache.spark.rdd.RDD
-import org.bdgenomics.adam.models.{
-  ReferenceRegion,
-  SequenceDictionary,
-  SequenceRecord
-}
+import org.bdgenomics.adam.models.{ReferenceRegion, SequenceDictionary, SequenceRecord}
+import org.bdgenomics.adam.rdd.OuterRegionJoinSuite
 import org.bdgenomics.adam.rdd.read.AlignmentRecordRDD
-import org.bdgenomics.adam.rdd.settheory.RightOuterShuffleRegionJoin
 import org.bdgenomics.formats.avro.AlignmentRecord
 
-class RightOuterShuffleRegionJoinSuite(partitionMap: Seq[Option[(ReferenceRegion, ReferenceRegion)]])
+class LeftOuterShuffleRegionJoinSuite(partitionMap: Seq[Option[(ReferenceRegion, ReferenceRegion)]])
     extends OuterRegionJoinSuite {
 
   val partitionSize = 3
@@ -41,6 +37,7 @@ class RightOuterShuffleRegionJoinSuite(partitionMap: Seq[Option[(ReferenceRegion
 
   def runJoin(leftRdd: AlignmentRecordRDD,
               rightRdd: AlignmentRecordRDD): RDD[(Option[AlignmentRecord], AlignmentRecord)] = {
-    RightOuterShuffleRegionJoin(leftRdd, rightRdd).compute()
+    LeftOuterShuffleRegionJoin(rightRdd, leftRdd)
+      .compute().map(_.swap)
   }
 }
