+++ adam-core/src/test/scala/org/bdgenomics/adam/rdd/sets/RightOuterShuffleRegionJoinSuite.scala
@@ -15,19 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.rdd
+package org.bdgenomics.adam.rdd.sets
 
 import org.apache.spark.rdd.RDD
-import org.bdgenomics.adam.models.{
-  ReferenceRegion,
-  SequenceDictionary,
-  SequenceRecord
-}
+import org.bdgenomics.adam.models.{ReferenceRegion, SequenceDictionary, SequenceRecord}
+import org.bdgenomics.adam.rdd.OuterRegionJoinSuite
 import org.bdgenomics.adam.rdd.read.AlignmentRecordRDD
-import org.bdgenomics.adam.rdd.settheory.LeftOuterShuffleRegionJoin
 import org.bdgenomics.formats.avro.AlignmentRecord
 
-class LeftOuterShuffleRegionJoinSuite(partitionMap: Seq[Option[(ReferenceRegion, ReferenceRegion)]])
+class RightOuterShuffleRegionJoinSuite(partitionMap: Seq[Option[(ReferenceRegion, ReferenceRegion)]])
     extends OuterRegionJoinSuite {
 
   val partitionSize = 3
@@ -41,7 +37,6 @@ class LeftOuterShuffleRegionJoinSuite(partitionMap: Seq[Option[(ReferenceRegion,
 
   def runJoin(leftRdd: AlignmentRecordRDD,
               rightRdd: AlignmentRecordRDD): RDD[(Option[AlignmentRecord], AlignmentRecord)] = {
-    LeftOuterShuffleRegionJoin(rightRdd, leftRdd)
-      .compute().map(_.swap)
+    RightOuterShuffleRegionJoin(leftRdd, rightRdd).compute()
   }
 }
